requirements.txt
absl-py==0.11.0
asgiref==3.3.1
dill==0.3.4
Django==3.1.7
django-cors-headers==3.7.0
dj-database-url==0.5.0
django-db==0.0.7
djangorestframework==3.12.4
gunicorn==20.1.0
h5py==3.6.0
hyperlink==21.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
sqlparse==0.4.1
tzdata==2021.5
whitenoise==6.1.0

